use consistent array query key normalization in SSG helpers
* apply `getArrayQueryKey` to `pathAndInput` in SSG helpers
